+++ src/result-tsk-legacy/src/index.ts
@@ -1,0 +1,4 @@
+export { IResult } from "./Result.interface";
+export { Result } from "./Result";
+export { IResult as IResultT } from "./generic/Result.interface";
+export { Result as ResultT } from "./generic/Result";