+++ src/validator-tsk/CHANGELOG.md
@@ -122,3 +122,9 @@
 - Methods was changed to camelCase.
 - Readme was updated.
 - Packages was updated.
+
+## [2.0.1] — 2022-01-02
+
+### Updated
+
+- Types was refactoring